backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.1.0
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
python-telegram-bot==20.7
stripe==7.6.0
requests==2.31.0
python-dotenv==1.0.0
django-storages==1.14.2
whitenoise==6.6.0
gunicorn==21.2.0
django-unfold==0.21.0 